Move api_handler
